Fix week_end to return end-of-day
This changes week_start and week_end so they explicitly return the
beginning and end of day respectively.

This fixes the problem where you ask for all the statuses in a given
week, but it's not returning the stuff on Sunday because it was doing
a <= of the beginning of Sunday rather than the end of Sunday.