Attempt at more helful error page (#260)
If there's a HELP_FAQ_URL defined, this will show a link to that on
the error page. It also shows a note for people trying to log in with
github and having difficulties.

We can't make it better than this since the problem occurs in the
login flow and we can't easily insert a handler there. Further, the
500 handler doesn't have any interesting context information, so we
can't figure it out in the template.

* Fixes to error page based on pmac's suggestions