Add .select_related() calls to reduce DB back-and-forth
This doesn't completely eliminate the db back-and-forth. The home page
still does an auth_user and a status DB query for every status it
shows on the page. The auth_user one is to get the email address to
show the gravatar. The status one is to get any replies.

Despite that, this should improve performance.