Tweak cache settings for better memory usage
We're running out of memory in prod constantly. My current theory is
that our cache settings are such that we're using an increasing amount
of cache until we pass the threshold for memory allowance.

Assuming that's correct, limiting the caching by number of entries and
timeout should make that better.

This implements that.