Redo auth0 implementation
This redoes the entire auth0 implementation as a separate app with a possibility
of converting it to a standalone library.

This implements the id_token renewing middleware as well as a model, but there
are no tests for it, no migration and it's not hooked up. I'll get there after I
see what things look like.