Build settings infrastructure
This builds a simple-ish settings infrastructure that'll be easier to generate
reST docs from when/if we convert this to a library. Further, it better codifies
defaults into the app code rather than in the Standup settings code. It also
ties the setting definitions with its respective documentation.

The end result of this is that it should be easier to configure for users and
easier to maintain for us.