Fix the 0005_add_team_user_id migration
When we migrated the Standup data from the old db schema to the new one, we
skipped the 0001_initial migration because we already had a bunch of tables and
data. However, we needed to make changes to those tables.

This migration adds an "id" column to the team table, but it already has one
from the 0001_initial migration.

Given that this migration is only helpful when updating an old Standup db, I
commented out the code and turned it into a no-op.