Fix teams and users
Previously, we had a TeamUser m2m model and that made lots of things hard
including the managing teams in the Django admin.

This does a meticulous set of migrations to copy that data to an ephemeral
table, nix that m2m model and recreate the relationship using Django magic.