Revert futures to 3.1.1
django-pipeline builds their wheel using Python 2.7, so when you install the
wheel using Python 3.6, it (incorrectly) requires `futures>=2.1.3`. That was
fine, but futures 3.2.0 adds a "don't install me for Python 3" thing, so
using futures 3.2.0 doesn't work.

This reverts futures back to 3.1.1 for now and hopefully upgrade to a version
of django-pipeline that doesn't require it.