Nix security tests
These tests test something that's part of Django and not something
that we implemented in Standup. Further, the way it's implemented
in Django means in order to do these tests, we'd need to recreate
the middlewares because the settings values are getting cached.

Given that, I'm going to nix the tests.